+++ packages/itwin/tree-widget/src/test/trees/models-tree/Utils.test.ts
@@ -4,12 +4,22 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { expect } from "chai";
+import { join } from "path";
+import { IModel } from "@itwin/core-common";
 import { StandardNodeTypes } from "@itwin/presentation-common";
-import { addModelsTreeNodeItemIcons, createRuleset, createSearchRuleset } from "../../../components/trees/models-tree/Utils";
+import {
+  buildTestIModel,
+  HierarchyCacheMode,
+  initialize as initializePresentationTesting,
+  terminate as terminatePresentationTesting,
+} from "@itwin/presentation-testing";
+import { waitFor } from "@testing-library/react";
+import { addModelsTreeNodeItemIcons, createRuleset, createSearchRuleset, queryModelsForHeaderActions } from "../../../components/trees/models-tree/Utils";
+import { addModel, addPartition } from "../../IModelUtils";
 
+import type { RepositoryLinkProps } from "@itwin/core-common";
 import type { DelayLoadedTreeNodeItem } from "@itwin/components-react";
 import type { Node } from "@itwin/presentation-common";
-
 describe("createRuleset", () => {
   it("creates default ruleset", () => {
     expect(createRuleset({})).to.matchSnapshot();
@@ -71,5 +81,54 @@ describe("addModelsTreeNodeItemIcons", () => {
     };
     addModelsTreeNodeItemIcons(item, node);
     expect(item.icon).to.be.eq("test-group-icon");
+  });
+});
+
+describe("queryModelsForHeaderActions", () => {
+  beforeEach(async () => {
+    await initializePresentationTesting({
+      backendProps: {
+        caching: {
+          hierarchies: {
+            mode: HierarchyCacheMode.Memory,
+          },
+        },
+      },
+      testOutputDir: join(__dirname, "output"),
+      backendHostProps: {
+        cacheDir: join(__dirname, "cache"),
+      },
+    });
+  });
+
+  afterEach(async () => {
+    await terminatePresentationTesting();
+  });
+
+  it("return zero available models when modeledElement is not GeometricElement3d or InformationPartitionElement", async () => {
+    // eslint-disable-next-line deprecation/deprecation
+    const iModel = await buildTestIModel("test", async (builder) => {
+      const repoLinkId = builder.insertElement({
+        model: IModel.repositoryModelId,
+        classFullName: "BisCore:RepositoryLink",
+        url: "url",
+        userLabel: "furl",
+      } as RepositoryLinkProps);
+
+      addModel(builder, "BisCore:PhysicalModel", repoLinkId);
+    });
+
+    const availableModels = await queryModelsForHeaderActions(iModel);
+    await waitFor(() => expect(availableModels.length).to.be.equal(0));
+  });
+
+  it("returns available model when modeled element is GeometricElement3d or InformationPartitionElement", async () => {
+    // eslint-disable-next-line deprecation/deprecation
+    const iModel = await buildTestIModel("test", async (builder) => {
+      const partition = addPartition(builder, "BisCore:PhysicalPartition", "partition");
+      addModel(builder, "BisCore:PhysicalModel", partition);
+    });
+    const availableModels = await queryModelsForHeaderActions(iModel);
+    await waitFor(() => expect(availableModels.length).to.be.equal(1));
   });
 });
